+++ Sources/TikTokOpenSDKCore/Internal/TikTokURLOpener.swift
@@ -11,6 +11,7 @@ import UIKit
 public protocol TikTokURLOpener {
     func canOpenURL(_ url: URL) -> Bool
     func open(_ url: URL, options: [UIApplication.OpenExternalURLOptionsKey : Any], completionHandler completion: ((Bool) -> Void)?)
+    var keyWindow: UIWindow? { get }
     func isTikTokInstalled() -> Bool
 }